+++ lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:domain_visualiser/widgets/app_widget/app_widget.dart';
+import 'package:domain_visualiser/widgets/app-init/app_widget.dart';
 import 'package:flutter/material.dart';
 
 void main() => runApp(AppWidget());
