+++ lib/widgets/app-init/app_widget.dart
@@ -9,8 +9,8 @@ import 'package:domain_visualiser/models/navigation/page_data/page_data.dart';
 import 'package:domain_visualiser/models/settings/settings.dart';
 import 'package:domain_visualiser/services/wrappers/firebase_wrapper.dart';
 import 'package:domain_visualiser/utils/redux_bundle.dart';
-import 'package:domain_visualiser/widgets/app_widget/initializing_error_page.dart';
-import 'package:domain_visualiser/widgets/app_widget/initializing_indicator.dart';
+import 'package:domain_visualiser/widgets/app-init/initializing_error_page.dart';
+import 'package:domain_visualiser/widgets/app-init/initializing_indicator.dart';
 import 'package:fast_immutable_collections/fast_immutable_collections.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
