--- lib/widgets/initial_page_builder.dart
@@ -1,22 +1,0 @@
-import 'package:domain_visualiser/widgets/auth/auth_page.dart';
-import 'package:domain_visualiser/widgets/drawing_page.dart';
-import 'package:domain_visualiser/widgets/error_page.dart';
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/material.dart';
-
-class InitialPageBuilder extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return StreamBuilder(
-        stream: FirebaseAuth.instance.authStateChanges(),
-        builder: (context, AsyncSnapshot<User?> snapshot) {
-          if (snapshot.hasError) {
-            return ErrorPage(error: snapshot.error);
-          }
-          if (snapshot.hasData) {
-            return (snapshot.data == null) ? AuthPage() : DrawingPage();
-          }
-          return CircularProgressIndicator();
-        });
-  }
-}
