+++ lib/widgets/shared/error_page.dart
